ISSUE-191 Client: Moved ChallengeMultipurposeBar up a directory.
