+++ client/src/components/challenge/main/response/ChallengeMultipurposeBar.js
